requirements.txt
asn1crypto==1.5.1
black==23.1.0
boto3==1.26.69
botocore==1.29.69
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.1
filelock==3.9.0
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
numpy==1.24.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
snowflake-connector-python==3.0.0
snowflake-sqlalchemy==1.4.6
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
